create constants and utils, fix burger bug
